fix errors and recognize position direction

+++ Colina/src/Colina.Language.CoreNLP/Analysers/PartOfSpeechAnalyser.cs
@@ -50,7 +50,7 @@ namespace Colina.Language.CoreNLP.Analysers
 
                 case "CD":
                     {
-                        userAction.ChangeRelativePosition(double.Parse(word), default(Direction));
+                        userAction.ChangeRelativePosition(int.Parse(word), default(Direction));
                         break;
                     }
